+++ invisibleroads_records/__init__.py
@@ -11,12 +11,12 @@ def includeme(config):
 
 def configure_settings(config):
     settings = config.get_settings()
-    set_default(
-        settings, 'sqlalchemy.url',
-        'sqlite:///%s/database.sqlite' % settings['data.folder'])
     for class_name, Class in CLASS_REGISTRY.items():
         if class_name.startswith('_'):
             continue
         key = Class.__tablename__ + '.id.length'
         value = set_default(settings, key, RECORD_ID_LENGTH, int)
         setattr(Class, 'id_length', value)
+    set_default(
+        settings, 'sqlalchemy.url',
+        'sqlite:///%s/database.sqlite' % settings['data.folder'])