+++ invisibleroads_records/scripts.py
@@ -1,6 +1,5 @@
 from invisibleroads.scripts import ConfigurableScript
-from invisibleroads_posts.routines.configuration import (
-    load_bootstrapped_settings)
+from invisibleroads_posts.routines.configuration import load_filled_settings
 
 from .models import Base, get_database_engine
 
@@ -10,7 +9,7 @@ class InitializeRecordsScript(ConfigurableScript):
     priority = 20
 
     def run(self, args):
-        settings = load_bootstrapped_settings(args.configuration_path)
+        settings = load_filled_settings(args.configuration_path)
         if 'sqlalchemy.url' not in settings:
             return
         database_engine = get_database_engine(settings)