+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 0.5
-- Remove sqlalchemy cache support
+- Remove sqlalchemy cache
+- Support database extensions
 
 # 0.2
 - Use server_cache.sqlalchemy setting prefix